User and user roles

+++ VagtPlan.ApiService/Models/WorkTime.cs
@@ -4,7 +4,7 @@ namespace ApiService.Models
 {
     public class WorkTime : Common
     {
-        public required int DepartmentId { get; set; }
+        public required long DepartmentId { get; set; }
         public required TimeOnly StartTime { get; set; }
         public required TimeOnly EndTime { get; set; }
         public required string DayOfWeek { get; set; }
